+++ CHANGELOG.md
@@ -60,6 +60,8 @@
   [#4102](https://github.com/Kong/kong-operator/pull/4102)
 - Hybridgateway: add support for `konghq.com/write-timeout` service annotation
   [#4103](https://github.com/Kong/kong-operator/pull/4103)
+- Hybridgateway: add support for `konghq.com/retries` service annotation
+  [#4104](https://github.com/Kong/kong-operator/pull/4104)
 - Add the following headers in requests of `ingress-controller` sent to Konnect
   for uploading configuration for tracing:
   - `X-Kic-Konnect-Sync-Instance-Id` for instance ID of Konnect config synchronizer.